pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "convert-nanoscope"
version = "0.0.1"

authors = [{name = "Constantine Evans", email = "const@costi.eu"}]

requires-python = ">=3.9"

dependencies = [
    "numpy >= 1",
    "matplotlib >= 3.6",
    "pillow >= 9.4.0",
    "tqdm >= 4.65",
    "click >= 8"   
]

[project.scripts]
convert-nanoscope = "convert_nanoscope.image:main"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]